pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "ruff>=0.12.10",
]
prod = [
    "gunicorn>=23.0.0",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

